+++ task/lib/src/widgets/add_task_dialog.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 import 'package:uuid/uuid.dart';
 
-import 'package:taskc/taskc.dart';
+import 'package:taskw/json.dart';
 
 import 'package:task/task.dart';