+++ taskw/lib/json.dart
@@ -1,0 +1,6 @@
+/// Custom JSON objects for various libraries of this package are organized
+/// here.
+
+library json;
+
+export 'src/json/task.dart';