+++ taskw/lib/taskc_impl.dart
@@ -1,0 +1,7 @@
+/// Implementation details for [taskc] library.
+
+library taskc_impl;
+
+export 'src/taskc/impl/codec.dart';
+export 'src/taskc/impl/message.dart';
+export 'src/taskc/impl/send.dart';