+++ taskw/lib/taskd.dart
@@ -1,0 +1,9 @@
+/// Helper functions used for setting up an installed
+/// [taskd](https://github.com/GothenburgBitFactory/taskserver), which in turn
+/// would be used for interactive development, or running tests locally or in
+/// CI.
+
+library taskd;
+
+export 'src/taskd/taskd.dart';
+export 'src/taskd/taskwarrior.dart';