+++ taskw/lib/taskrc.dart
@@ -1,0 +1,7 @@
+/// Currently this library only contains the parseTaskrc function, which is
+/// currently only used for configuring a client to point at a
+/// [Taskserver](https://github.com/GothenburgBitFactory/taskserver).
+
+library taskrc;
+
+export 'src/taskrc/taskrc.dart';