+++ taskw/CHANGELOG.md
@@ -1,1 +1,4 @@
 # 0.0.0-alpha.0
+
+- Merge libraries taskd-client and taskd-setup into the taskw-dart
+  package.